+++ webapp/channels/package.json
@@ -103,6 +103,7 @@
     "zen-observable": "0.9.0"
   },
   "devDependencies": {
+    "@calls/common": "github:mattermost/calls-common#v0.26.0",
     "@deanwhillier/jest-matchmedia-mock": "1.2.0",
     "@hot-loader/react-dom": "17.0.2",
     "@mattermost/eslint-plugin": "*",
