+++ webapp/channels/src/selectors/calls.ts
@@ -1,11 +1,14 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import type {CallsConfig, UserSessionState} from '@calls/common/lib/types';
 import semver from 'semver';
 
 import {suitePluginIds} from 'utils/constants';
 
 import type {GlobalState} from 'types/store';
+
+const CALLS_PLUGIN = 'plugins-com.mattermost.calls';
 
 export function isCallsEnabled(state: GlobalState, minVersion = '0.4.2') {
     return Boolean(state.plugins.plugins[suitePluginIds.calls] &&
@@ -17,4 +20,16 @@ export function isCallsRingingEnabledOnServer(state: GlobalState) {
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     return Boolean(state[`plugins-${suitePluginIds.calls}`]?.callsConfig?.EnableRinging);
+}
+
+export function getSessionsInCalls(state: GlobalState): Record<string, Record<string, UserSessionState>> {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    return state[CALLS_PLUGIN].sessions || {};
+}
+
+export function getCallsConfig(state: GlobalState): CallsConfig {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    return state[CALLS_PLUGIN].callsConfig;
 }
