+++ webapp/package.json
@@ -26,7 +26,6 @@
     "@babel/preset-env": "7.21.5",
     "@babel/preset-react": "7.18.6",
     "@babel/preset-typescript": "7.21.5",
-    "@calls/common": "github:mattermost/calls-common#v0.26.0",
     "@types/webpack-env": "1.18.1",
     "babel-loader": "9.1.2",
     "babel-plugin-formatjs": "10.5.1",
